adjlist, Dij, Shortest path on tree (LCA), DSU
This problem requires the participant to have a comprehensive coding & debugging ability, for it requires: 
Adjacent list, Dijkstra, Shortest path on tree (which includes LCA) and DSU
The idea to solve the problem is quite easy, which depends on the key observation that m - n <= 20. We can get shortest path
of the spanning tree, and use Dijkstra to bruteforce the shortest path between queried points and the points which are involved
in the redundant edges.

The problem took me a lot of time to debug, but at last I found my problem was just that I forgot to initialize the lg array!
Need to be more careful next time.